feat(Logger): Add improvements to log
