+++ ui/src/components/settings/sub_pages/mod.rs
@@ -3,6 +3,5 @@ pub mod developer;
 pub mod extensions;
 pub mod files;
 pub mod general;
-pub mod logger_debug;
 pub mod privacy;
 pub mod profile;
