+++ ui/src/logger/style.scss
@@ -1,0 +1,13 @@
+.logger-body {
+    background-color: black;
+    height: 100%;
+    margin: 0;
+    overscroll-behavior-y: none;
+    overscroll-behavior-x: none;
+    overflow: hidden;
+}
+
+.log-text {
+    font-size: 12px;
+    margin-left: 16px;
+}
